Adaugare identificator server :stars:

+++ MTADroneService_server/DroneService/src/main/java/MTADroneService/DroneService/application/services/implementation/ServiceServiceImpl.java
@@ -1,22 +1,20 @@
-package MTADroneService.DroneService.authentification.services.implementation;
+package MTADroneService.DroneService.application.services.implementation;
 
-import MTADroneService.DroneService.authentification.daos.DroneDAO;
-import MTADroneService.DroneService.authentification.daos.MissionDAO;
-import MTADroneService.DroneService.authentification.dtos.MissionInfoDTO;
-import MTADroneService.DroneService.authentification.models.DroneModel;
-import MTADroneService.DroneService.authentification.models.MissionModel;
-import MTADroneService.DroneService.authentification.services.ServiceService;
-import MTADroneService.DroneService.authentification.utility.Client;
-import MTADroneService.DroneService.authentification.utility.SSHConnection;
-import MTADroneService.DroneService.authentification.utility.Utils;
-import MTADroneService.DroneService.authentification.utility.implementation.ClientImpl;
+import MTADroneService.DroneService.application.daos.DroneDAO;
+import MTADroneService.DroneService.application.daos.MissionDAO;
+import MTADroneService.DroneService.application.daos.UserDAO;
+import MTADroneService.DroneService.application.dtos.MissionInfoDTO;
+import MTADroneService.DroneService.application.models.DroneModel;
+import MTADroneService.DroneService.application.models.MissionModel;
+import MTADroneService.DroneService.application.models.UserModel;
+import MTADroneService.DroneService.application.services.ServiceService;
+import MTADroneService.DroneService.application.utility.DroneCoordinates;
+import MTADroneService.DroneService.application.utility.MissionInfoToSend;
+import MTADroneService.DroneService.application.utility.Utils;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.awt.*;
-import java.io.File;
 import java.io.IOException;
 import java.util.*;
 import java.util.List;
@@ -33,23 +31,32 @@ public class ServiceServiceImpl implements ServiceService {
     @Autowired
     DroneDAO droneDAO;
 
+    @Autowired
+    UserDAO userDAO;
+
     @Override
-    public void createMission(MissionInfoDTO missionInfoDTO) throws IOException {
+    public void createMission(MissionInfoDTO missionInfoDTO, String jwtToken,
+                              MissionInfoToSend missionInfoToSend) throws IOException {
 
         MissionModel missionModel = modelMapper.map(missionInfoDTO, MissionModel.class);
-        missionModel.setMissionID(UUID.randomUUID().toString());
         missionModel.setMissionStatus(Utils.MissionStatus.PREPARING.toString());
         List<DroneModel> droneModelList = droneDAO.findByDroneStatus("AVAILABLE");
 
-        if(droneModelList.size() != 0) {
-            DroneModel newDrone = droneModelList.get(0);
-            newDrone.setDroneStatus(Utils.DroneStatus.BUSY.toString());
-            droneDAO.save(newDrone);
-            droneDAO.deleteByDroneIDAndDroneStatus(droneModelList.get(0).getDroneID(), Utils.DroneStatus.AVAILABLE.toString());
-            missionModel.setMissionDroneID(newDrone.getDroneID());
-            missionDAO.save(missionModel);
-            SSHConnection sshConnection = new SSHConnection("172.20.10.4","pi","raspberry");
-            sshConnection.sendCommandToDrone("sudo python /home/pi/MTADroneService/drone_controller/drone_client.py");
+        Optional<UserModel> optionalUserModel = userDAO.findByJwtToken(jwtToken);
+        if(optionalUserModel.isPresent()){
+            UserModel userModel = optionalUserModel.get();
+            missionModel.setUserID(userModel.getUserID());
+            if(droneModelList.size() != 0) {
+                DroneModel newDrone = droneModelList.get(0);
+                Utils.addMissionInfoToSend(newDrone.getDroneID(),missionInfoToSend);
+                newDrone.setDroneStatus(Utils.DroneStatus.BUSY.toString());
+                droneDAO.save(newDrone);
+                droneDAO.deleteByDroneIDAndDroneStatus(droneModelList.get(0).getDroneID(), Utils.DroneStatus.AVAILABLE.toString());
+                missionModel.setMissionDroneID(newDrone.getDroneID());
+                missionDAO.save(missionModel);
+                DroneCoordinates droneCoordinates = new DroneCoordinates();
+                Utils.currentPosition.put(newDrone.getDroneID(),droneCoordinates);
+            }
         }
     }
 }
